requirements.txt
aiogram==2.17.1
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.9
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
sqlalchemy==1.4.28
psycopg2-binary==2.9.2
greenlet==1.1.2
asyncpg==0.25.0
alembic==1.7.6